backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyPDF2==3.0.1
faiss-cpu==1.8.0
openai>=1.50.0,<2.0.0
tiktoken==0.7.0
python-dotenv==1.0.1
httpx==0.27.2
